python-backend/requirements.txt
# FastAPI Framework and Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# PII Detection (Microsoft Presidio)
presidio-analyzer>=2.2.354
presidio-anonymizer>=2.2.354

# Document Parsing and AI
llama-parse>=0.3.0
anthropic>=0.18.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9

# Utilities
python-dotenv>=1.0.0
pydantic>=2.6.0

# Additional Dependencies
pytest>=7.4.0              # For testing
httpx>=0.26.0              # For async HTTP requests
cryptography>=42.0.0       # For PII encryption (Phase 2)
python-multipart>=0.0.6    # For file uploads
